게시판 수정 시 이미지파일수정 적용 완료

+++ ttamna/src/main/java/com/kh/ttamna/repository/donation/DonationDaoImpl.java
@@ -82,10 +82,11 @@ public class DonationDaoImpl implements DonationDao{
 	}
 	
 	@Override//기부 후 금액 업데이트
-	public boolean funding(int donationNo, int price) {
+	public boolean funding(int donationNo, long price) {
 		Map<String, Object> map = new HashMap<>();
 		map.put("donationNo", donationNo);
 		map.put("price", price);
 		return sqlSession.update("donation.funding", map) > 0;
 	}
+	
 }
